Modularize and restructured runInContext
- in anticipation for fullJS and Source Native mode
